autobl/autobl/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "autobl"
authors = [
    {name = ""},
]
description = "Automation of beamline operations"
readme = "README.md"
license = {file = "LICENSE.txt"}
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "scipy",
    "matplotlib",
    "tqdm",
    "jupyter",
    "ipywidgets",
    "scikit-learn",
    "scikit-image",
    "pandas"
]
version = "0.0.1"

[tool.setuptools.packages]
find = {}  # Scan the project directory with the default parameters

[project.optional-dependencies]
steering = ["botorch == 0.9.4"]
